requirements.txt
torch==2.6.0
torchvision
transformers==4.46.3
tokenizers==0.20.3
safetensors>=0.4.0
accelerate>=0.26.0
protobuf>=3.20.0
spaces>=0.20.0
Pillow>=10.0.0
einops
addict 
easydict
gradio
einops
addict 
easydict
# pip install flash-attn==2.7.3 --no-build-isolation  (only support 3xxx or higher RTX GPU with Ampere architecture)